[Maps] refactor static can skip logic from layer class (#51627)
